+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weaverse/core
 
+## 5.2.0
+
+### Minor Changes
+
+- update type system with new schema validation util
+
 ## 5.1.0
 
 ### Minor Changes