+++ packages/shopify/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weaverse/shopify
 
+## 5.2.0
+
+### Patch Changes
+
+- @weaverse/react@5.2.0
+
 ## 5.1.0
 
 ### Minor Changes